requirements.txt
streamlit~=1.39.0
transformers~=4.46.0
PyPDF2~=3.0.1
google-generativeai
protobuf~=5.28.3
torch~=2.5.0